# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "📚 Biblioteca REST API",
        "description": "API desarrollada por **Jenifer Alvarez** (Sprint 5 - IT Academy).  \r\n *\r\n * Implementa autenticación con **Laravel Passport**, control de acceso por roles  \r\n * y gestión de **Libros**, **Categorías** y **Préstamos**.  \r\n * Respuestas JSON estructuradas para integrarse con frontends o sistemas de IA.",
        "contact": {
            "name": "Jenifer Álvarez",
            "url": "https://github.com/jenifera5"
        },
        "license": {
            "name": "MIT License",
            "url": "https://opensource.org/licenses/MIT"
        },
        "version": "1.0.0"
    },
    "servers": [
        {
            "url": "http://127.0.0.1:8000",
            "description": "Servidor local de desarrollo"
        }
    ],
    "paths": {
...
        "/loans": {
            "get": {
                "tags": [
                    "Préstamos"
                ],
                "summary": "Listar todos los préstamos",
                "operationId": "766507ce16ccc64c1d52008c7dd80908",
                "responses": {
                    "200": {
                        "description": "Lista de préstamos obtenida correctamente",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "type": "array",
                                    "items": {
                                        "properties": {
                                            "id": {
                                                "type": "integer",
                                                "example": 1
                                            },
                                            "id_usuario": {
                                                "type": "integer",
                                                "example": 3
                                            },
                                            "id_libro": {
                                                "type": "integer",
                                                "example": 5
                                            },
                                            "fecha_prestamo": {
                                                "type": "string",
                                                "format": "date",
                                                "example": "2025-10-10"
                                            },
                                            "fecha_devolucion": {
                                                "type": "string",
                                                "format": "date",
                                                "example": "2025-10-25"
                                            },
                                            "estado": {
                                                "type": "string",
                                                "example": "pendiente"
                                            },
                                            "usuario": {
                                                "properties": {
                                                    "nombre": {
                                                        "type": "string",
                                                        "example": "Jenifer"
                                                    }
                                                },
                                                "type": "object"
                                            },
                                            "libro": {
                                                "properties": {
                                                    "titulo": {
                                                        "type": "string",
                                                        "example": "Dune"
                                                    }
                                                },
                                                "type": "object"
                                            }
                                        },
                                        "type": "object"
                                    }
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "passport": []
                    }
                ]
            },
            "post": {
                "tags": [
                    "Préstamos"
                ],
                "summary": "Crear un nuevo préstamo",
                "operationId": "43f3f65f079f232d18323322efa77095",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "id_usuario",
                                    "id_libro",
                                    "fecha_prestamo",
                                    "estado"
                                ],
                                "properties": {
                                    "id_usuario": {
                                        "type": "integer",
                                        "example": 1
                                    },
                                    "id_libro": {
                                        "type": "integer",
                                        "example": 3
                                    },
                                    "fecha_prestamo": {
                                        "type": "string",
                                        "format": "date",
                                        "example": "2025-10-10"
                                    },
                                    "fecha_devolucion": {
                                        "type": "string",
                                        "format": "date",
                                        "example": "2025-10-20"
                                    },
                                    "estado": {
                                        "type": "string",
                                        "example": "pendiente"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "Préstamo creado correctamente",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "message": {
                                            "type": "string",
                                            "example": "Prestamo creado correctamente"
                                        },
                                        "prestamo": {
                                            "properties": {
                                                "id": {
                                                    "type": "integer",
                                                    "example": 2
                                                },
                                                "id_usuario": {
                                                    "type": "integer",
                                                    "example": 1
                                                },
                                                "id_libro": {
                                                    "type": "integer",
                                                    "example": 3
                                                },
                                                "fecha_prestamo": {
                                                    "type": "string",
                                                    "format": "date",
                                                    "example": "2025-10-10"
                                                },
                                                "fecha_devolucion": {
                                                    "type": "string",
                                                    "format": "date",
                                                    "example": "2025-10-20"
                                                },
                                                "estado": {
                                                    "type": "string",
                                                    "example": "pendiente"
                                                }
                                            },
                                            "type": "object"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "422": {
                        "description": "Datos inválidos"
                    }
                },
                "security": [
                    {
                        "passport": []
                    }
                ]
            }
        },
        "/loans/{id}": {
            "put": {
                "tags": [
                    "Préstamos"
                ],
                "summary": "Actualizar un préstamo existente",
                "operationId": "56552c7aa6b966989a00c5f7d484db15",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "ID del préstamo a actualizar",
                        "required": true,
                        "schema": {
                            "type": "integer",
                            "example": 4
                        }
                    }
                ],
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "id_usuario",
                                    "id_libro",
                                    "fecha_prestamo",
                                    "estado"
                                ],
                                "properties": {
                                    "id_usuario": {
                                        "type": "integer",
                                        "example": 2
                                    },
                                    "id_libro": {
                                        "type": "integer",
                                        "example": 6
                                    },
                                    "fecha_prestamo": {
                                        "type": "string",
                                        "format": "date",
                                        "example": "2025-10-01"
                                    },
                                    "fecha_devolucion": {
                                        "type": "string",
                                        "format": "date",
                                        "example": "2025-10-15"
                                    },
                                    "estado": {
                                        "type": "string",
                                        "example": "devuelto"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Préstamo actualizado correctamente",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "message": {
                                            "type": "string",
                                            "example": "Prestamo actualizado correctamente"
                                        },
                                        "prestamo": {
                                            "properties": {
                                                "id": {
                                                    "type": "integer",
                                                    "example": 4
                                                },
                                                "id_usuario": {
                                                    "type": "integer",
                                                    "example": 2
                                                },
                                                "id_libro": {
                                                    "type": "integer",
                                                    "example": 6
                                                },
                                                "fecha_prestamo": {
                                                    "type": "string",
                                                    "format": "date",
                                                    "example": "2025-10-01"
                                                },
                                                "fecha_devolucion": {
                                                    "type": "string",
                                                    "format": "date",
                                                    "example": "2025-10-15"
                                                },
                                                "estado": {
                                                    "type": "string",
                                                    "example": "devuelto"
                                                }
                                            },
                                            "type": "object"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "Préstamo no encontrado"
                    }
                },
                "security": [
                    {
                        "passport": []
                    }
                ]
            },
            "delete": {
                "tags": [
                    "Préstamos"
                ],
                "summary": "Eliminar un préstamo",
                "operationId": "d28a4691e1f1057d817a9496a56e0fa7",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "ID del préstamo a eliminar",
                        "required": true,
                        "schema": {
                            "type": "integer",
                            "example": 7
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Préstamo eliminado correctamente",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "message": {
                                            "type": "string",
                                            "example": "Prestamo eliminado correctamente"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "Préstamo no encontrado"
                    }
                },
                "security": [
                    {
                        "passport": []
                    }
                ]
            }
        }
    },
    "components": {
        "securitySchemes": {
            "bearerAuth": {
                "type": "http",
                "bearerFormat": "JWT",
                "scheme": "bearer"
            }
        }
    },
...
}